+++ UniTAS/Plugin/Services/IMonoBehEventInvoker.cs
@@ -4,4 +4,6 @@ public interface IMonoBehEventInvoker
 {
     void Update();
     void FixedUpdate();
+    void LateUpdate();
+    void OnGUI();
 }